Various small changes

+++ app/src/main/java/bg/o/sim/finansizmus/model/Transaction.java
@@ -91,7 +91,7 @@ public class Transaction implements Serializable {
 
         @Override
         public int compare(Transaction o1, Transaction o2) {
-            if (o1.date.equals(o2.date)) return ((Double) o2.sum).compareTo((Double) o1.sum);
+            if (o1.date.equals(o2.date)) return ((Double) o2.sum).compareTo(o1.sum);
             return o2.date.compareTo(o1.date);
         }
 
@@ -106,7 +106,7 @@ public class Transaction implements Serializable {
         @Override
         public int compare(Transaction o1, Transaction o2) {
             if (o1.sum == o2.sum) return o2.date.compareTo(o1.date);
-            return ((Double) o2.sum).compareTo((Double) o1.sum);
+            return ((Double) o2.sum).compareTo(o1.sum);
         }
 
         @Override
@@ -121,7 +121,7 @@ public class Transaction implements Serializable {
         public int compare(Transaction o1, Transaction o2) {
             if (o1.category.equals(o2.category)) {
                 if (o1.date.equals(o2.date))
-                    return ((Double) o2.sum).compareTo((Double) o1.sum);
+                    return ((Double) o2.sum).compareTo(o1.sum);
                 return o2.date.compareTo(o1.date);
             }
